captcha: Fold the captcha into the registration forms themselves.
The two captcha-enabled registration forms were identical apart
from their base class, and existed alongside captcha-less versions
of the same forms.  That pairing was a hazard: nothing stops new
code from constructing the plain form where the captcha-enabled one
was intended, and since captchas are disabled in development and
tests, the missing protection would be invisible until it hit
production.

Extract the shared field, validation, and constructor into a mixin
on the registration forms themselves, which removes its own captcha
field when the server has no captcha configured.  The views always
construct the same form class, and a captcha-less variant no longer
exists to reach for by accident.
